Extract quota details into new class
