+++ src/main/java/com/emc/ecs/managementClient/model/BucketQuotaDetails.java
@@ -3,10 +3,8 @@ package com.emc.ecs.managementClient.model;
 import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name = "bucket_quota_details")
-public class BucketQuotaDetails {
+public class BucketQuotaDetails extends QuotaDetails {
 
-	private long blockSize;
-	private long notificationSize;
 	private String bucketname;
 	private String namespace;
 
@@ -24,21 +22,5 @@ public class BucketQuotaDetails {
 
 	public void setNamespace(String namespace) {
 		this.namespace = namespace;
-	}
-	
-	public long getBlockSize() {
-		return blockSize;
-	}
-
-	public void setBlockSize(long blockSize) {
-		this.blockSize = blockSize;
-	}
-
-	public long getNotificationSize() {
-		return notificationSize;
-	}
-
-	public void setNotificationSize(long notificationSize) {
-		this.notificationSize = notificationSize;
 	}
 }
